fix: use single quotes in exports
This fixes an edge case where the value of an exported variable can
contain command substitutions. For example, it is common to have PR
templates in GitHub, and if the template contains any backticks (`), the
DRONE_COMMIT_MESSAGE will contain command substitutions. In my case,
this caused our builds to fail because we give example make commands,
which are real targets, and this is unexpected consequences.

Signed-off-by: Andrew Rynhard <andrew@rynhard.io>